+++ CHANGELOG.md
@@ -60,6 +60,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Add SDL validation rule `UniqueTypeNames` (#998)
 - Add support for SDL validation to `KnownTypeNames` rule (#999)
 - Add SDL validation rule `UniqueArgumentDefinitionNames` (#1136)
+- Add `parseValue` config option to InputObjectType to parse input value to custom value object
 - Add option `sortTypes` to have `SchemaPrinter` order types alphabetically
 
 ### Optimized